Move permissions and constants modules
